Implement predefined assembly workcell and resource grounding tests
- Added `workcell.py` to define the PredefinedWorkcellSnapshot and related functionality for validating assembly capabilities.
- Created tests for the predefined workcell in `test_predefined_workcell.py` to ensure it contains only the expected resources and validates against changes.
- Introduced `test_pa_resource_grounding.py` to validate resource assignment and selection logic based on the predefined workcell.
- Developed `test_resource_registry.py` to ensure the resource registry correctly projects resource identities and handles validation errors.
- Implemented `gazebo_observation_worker.py` for capturing observations in a ROS2 environment, handling errors and writing results to a JSON file.

+++ cais_spade_llm/spec2primitives/tools/rgb_d_cad_grounding/segmenter.py
@@ -476,7 +476,7 @@ def _segment_camera(
                     "minimum": _float_list(candidate_points.min(axis=0)),
                     "maximum": _float_list(candidate_points.max(axis=0)),
                 },
-                "centroid_m": _float_list(candidate_points.mean(axis=0)),
+                "centroid_m": _float_list(candidate_points.mean(axis=0, dtype=np.float64)),
                 "depth_range_m": {
                     "minimum": float(candidate_points[:, 2].min()),
                     "maximum": float(candidate_points[:, 2].max()),
